Distribution Selection with AIC Criteria

Performs a maximum likelihood estimation for univariate and multivariate generalized hyperbolic distributions and reorder the fitted models based on AIC information criteria.

Usage

reveal(.invariant, .silent = TRUE, ...)

# S3 method for default
reveal(.invariant, .silent = TRUE, ...)

# S3 method for tbl_df
reveal(.invariant, .silent = TRUE, ...)

# S3 method for xts
reveal(.invariant, .silent = TRUE, ...)

# S3 method for matrix
reveal(.invariant, .silent = TRUE, ...)

# S3 method for numeric
reveal(.invariant, .silent = TRUE, ...)

Arguments

.invariant

A tabular (non-tidy) time-series.

.silent

A logical. Should the optimization results be printed? The default is TRUE.

...

Additional arguments the user may want to pass to fit.ghypuv.

Value

A tibble with the fitted parameters: 11 rows and 8 columns.

References

Pfaff, Bernhard. Financial risk modelling and portfolio optimization with R. John Wiley & Sons, 2016.

See also

Examples

x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- colnames(EuStockMarkets)

reveal(x)
#> Currently fitting: asymmetric ghyp 
#> Currently fitting: asymmetric hyp 
#> Currently fitting: asymmetric NIG 
#> Currently fitting: asymmetric VG 
#> Currently fitting: asymmetric t 
#> Currently fitting: symmetric ghyp 
#> Currently fitting: symmetric hyp 
#> Currently fitting: symmetric NIG 
#> Currently fitting: symmetric VG 
#> Currently fitting: symmetric t 
#> Currently fitting: gauss
#> # A tibble: 11 × 8
#>    model symmetric lambda  alpha.bar     aic    llh converged n.iter
#>    <chr> <lgl>      <dbl>      <dbl>   <dbl>  <dbl> <lgl>      <dbl>
#>  1 ghyp  FALSE      0.655   6.33e-14 -56122. 28081. TRUE          39
#>  2 ghyp  TRUE       0.654   4.35e-13 -55847. 27940. TRUE          39
#>  3 VG    FALSE      0.655   0        -54375. 27207. TRUE          23
#>  4 VG    TRUE       0.654   0        -54369. 27200. TRUE          24
#>  5 hyp   TRUE       2.5     2.58e- 7 -52721. 26375. TRUE           9
#>  6 hyp   FALSE      2.5     9.00e- 8 -52715. 26376. TRUE          18
#>  7 t     TRUE      -3.08    0        -52711. 26371. TRUE          63
#>  8 t     FALSE     -3.10    0        -52711. 26375. TRUE          64
#>  9 NIG   FALSE     -0.5     1.87e+ 0 -52708. 26373. TRUE          37
#> 10 NIG   TRUE      -0.5     1.85e+ 0 -52708. 26369. TRUE          36
#> 11 gauss TRUE      NA     Inf        -52096. 26062. TRUE           0